+++ src/main/resources/templates/back-end/forumPost/listOneForumPost.html
@@ -34,23 +34,21 @@
 
 <div align="center" th:fragment="listOneForumPost-div">
     <div th:if="${forumPostVO.fpStat == 1}">
-        <!-- 當 fpStat 為 1 時顯示的內容 -->
         <h3><span>查詢結果 :</span></h3>
         <span>貼文編號:</span><span th:text="${forumPostVO.fpNum}"></span><br/>
         <span>會員編號:</span><span th:text="${forumPostVO.userVO.userId} + '-[' + ${forumPostVO.userVO.comName} + ']'"></span><br/>
         <span>產業類別:</span><span th:text="${forumPostVO.fpCategory}"></span><br/>
         <span>文章標題:</span><span th:text="${forumPostVO.fpTitle}"></span><br/>
         <span>文章內容:</span><span th:text="${forumPostVO.fpContent}"></span><br/>
-        <span>貼文圖片:</span><img th:src="@{/forumPost/DBGifReader3(fpNum=${forumPostVO.fpNum})}" width="100px"><br/>
-        <span>貼文發佈時間:</span><span th:text="${forumPostVO.fpTime}"></span><br/>
-        <span>貼文最後編輯時間:</span><span th:text="${forumPostVO.fpUpdate}"></span><br/>
+        <span>貼文圖片:</span><img th:src="@{/forumPost/DBGifReader(fpNum=${forumPostVO.fpNum})}" width="100px"><br/>
+        <span>貼文發佈時間:</span><span th:text="${#dates.format(forumPostVO.fpTime, 'yyyy-MM-dd HH:mm')}"></span><br/>
+        <span>貼文最後編輯時間:</span><span th:text="${#dates.format(forumPostVO.fpUpdate, 'yyyy-MM-dd HH:mm')}"></span><br/>
         <span>貼文愛心數:</span><span th:text="${forumPostVO.fpLike}"></span><br/>
-<!--         <span>貼文狀態:</span><span th:text="${forumPostVO.fpStat}"></span><br/> -->
     </div>
     <div th:unless="${forumPostVO.fpStat == 1}">
-<!--         當 fpStat 不為 1 時顯示的內容 -->
+		<!--當fpStat 不為 1 時顯示的內容 -->
         <h3>此篇文章已刪除</h3>
-        在這裡插入圖片
+        <!-- 在這裡插入圖片 --> 
         <img width="90px"  height="80px" th:src="@{/images/delete.jpg}">
     </div>
 </div>
