+++ src/main/resources/templates/front-end/footer.html
@@ -11,6 +11,12 @@
 				<li class="nav-item"><a th:href="@{/userinformation/customer_service}"
 					class="nav-link px-2 text-body-secondary">聯絡我們</a></li>
 			</ul>
+			
+			 <div class="social-icons">
+              <i class="fa-regular fa-envelope fa-2xl" style="color: #dfd8d8; margin-right: 15px;"></i>
+              <i class="fa-brands fa-github fa-2xl" style="color: #dfd8d8;"></i>
+            </div>
+            
 			<p class="text-center text-body-secondary">&copy; 2023 NO BARRIER
 				TECH Company, Inc</p>
 		</footer>
